+++ trunk/TABU VRP/src/tabuvrp/core/move/Move.java
@@ -3,16 +3,16 @@ package tabuvrp.core.move;
 
 public abstract class Move {
 
-    private final int deltaCost;
+    private final double deltaCost;
     private final int deltaDemandBalance;
 
-    public Move(int deltaCost,
+    public Move(double deltaCost,
                 int deltaDemandBalance) {
         this.deltaCost = deltaCost;
         this.deltaDemandBalance = deltaDemandBalance;
     }
     
-    public final int getDeltaCost() {
+    public final double getDeltaCost() {
         return deltaCost;
     }
